+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changelog
 
+- Rebase on upstream (`1.5.1`) version (crds & all the manifests).
 - Changed hard-coded nodeSelector into configurable tolerations-affinity combo.
 
 ## [0.8.1] - 2021-10-25